[WTF] Share ProcessMemoryStatus declaration among Linux and Haiku https://bugs.webkit.org/show_bug.cgi?id=303774
Reviewed by Michael Catanzaro.

Add a new <wtf/ProcessMemoryStatus.h> with the definition of the
ProcessMemoryStatus struct, and then reuse that from both the Linux-
and Haiku-specific implementations. While at it, annotate the struct
with WTF_FORBID_HEAP_ALLOCATION because its usage is always stack
allocated, and ideally should stay that way.

Canonical link: https://commits.webkit.org/305590@main
